requirements.txt
python-telegram-bot==20.7
python-dotenv>=0.21.0,<1.0
requests>=2.28.0,<3.0
urllib3>=1.26.0,<2.0
pytz>=2023.3,<2024.0
aiohttp>=3.8.1,<4.0
python-dateutil>=2.8.2,<3.0
flask>=2.3.2,<3.0
